only apply mask to _led_buffer if < 256
This works on CPython.  I _think_ it will work on CircuitPython, since
the old behavior is effectively a no-op there.  On CPython it was
throwing a `ValueError`.

I want to be clear that I don't understand this code, so I'm definitely
not confident that this change is the Correct Thing.